Plan 9 from Bell Labs Fourth Edition 2007-03-04

+++ sys/src/cmd/ql/l.h
@@ -50,7 +50,7 @@ struct	Prog
 	short	line;
 	short	mark;
 	short	optab;		/* could be uchar */
-	uchar	as;
+	ushort	as;
 	char	reg;
 };
 struct	Sym
@@ -75,7 +75,7 @@ struct	Autom
 };
 struct	Optab
 {
-	uchar	as;
+	ushort	as;
 	char	a1;
 	char	a2;
 	char	a3;
